Check more range, refactor
I am adding more range checking functionality to check to see if the time range is in AM vs PM and if the range goes from evening to morning. I've also begun refactoring more to clean up the code. Getters and Setters were added to help clean up a lot of variable use as the number of checks increased.