Cleaned up some logic in testing
I was able to get Family A to work with more scenarios in testing. I still need to work on a few more. After this I will refactor again. I see some variables that may be a better fit in the base class so other extended classes can use them.