Family C Complete Family A refactor
finished testing Family C and made adjustments. As I was finishing family C I went back and looked at the logic of Family A which is a similar set up. I liked the way I wrote Family C better even though the results were the same (C is clearer and simpler). I went back and refactored family A using the logic for C. Testing for A did show that it still worked properly.

Next will be a finalized input validation and entry logic. Most checks are in place (but not enforced).