Add more checks in Main, worked on Family A
As I was implementing the validation checks I did find a number of places where data wasn't being set correctly. This also lead to other areas in the Family A class that produced an incorrect calculation. I made corrections for the case that failed. I will re-test other areas to double check what was missed and why it may have worked the first time.